chore: rename chat presenter
