+++ room-front/src/component/RoomPage/Chat.js
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import ChatAPi from '../../api/ChatAPi'
 import { COLOR } from '../../constant/style'
 import useSocket from '../../hooks/useSocket'
-import ChatPresenter from '../../presenter/chat/Chat'
+import ChatPresenter from '../../presenter/chat/ChatPresenter'
 import RoomHeader from '../../presenter/header/RoomHeader'
 import PageWrapper from '../../presenter/wrapper/PageWrapper'
 import ChatThread from './ChatThread'
